Create README.md
A student portal where each student can access the details of books issued by them and their due date. The best thing about this portal is that it sends a reminder email to the student a day prior to the due date so that now students can keep a track of the due dates of their books. 📚  Additionally, incase the student miss the due date of the book then the portal shows the fine amount they need to submit till date.